format update
add fontawesome, for email and linkedin icons

add bootstrap. Update the buttons to use bootstrap libraries

adds card-columns to all sections, so that everything is more responsive to the size of the screen. This saves real-estate on big screens, while still being readable on small screens.

add the education section